// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/DetailedAgreementDTO.ts
export const DetailedAgreementDTOProductTypeEnum = {
    Rbf: 'RBF',
    Fixed: 'FIXED',
    InterestRate: 'INTEREST_RATE',
    DailyPayout: 'DAILY_PAYOUT',
    Unknown: 'UNKNOWN'
} as const;
export type DetailedAgreementDTOProductTypeEnum = typeof DetailedAgreementDTOProductTypeEnum[keyof typeof DetailedAgreementDTOProductTypeEnum];