// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/Transaction.ts
export interface Transaction {
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    name: string;
    /**
     * 
     * @type {Date}
     * @memberof Transaction
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Transaction
     */
    operationScheduledDate?: Date;
    /**
     * Monetary value represented as a floating point number (the amount in agreement currency) with currency
     * @type {MonetaryAmount}
     * @memberof Transaction
     */
    transactionAmount: MonetaryAmount;
    /**
     * Monetary value represented as a floating point number (the amount + fees) with currency
     * @type {MonetaryAmount}
     * @memberof Transaction
     */
    totalAmount: MonetaryAmount;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    type: TransactionTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    status: TransactionStatusEnum;
    /**
     * 
     * @type {TransactionAccount}
     * @memberof Transaction
     */
    account: TransactionAccount;
}


/**
 * @export
 */
export const TransactionTypeEnum = {
    Invoice: 'INVOICE',
    Card: 'CARD',
    Cash: 'CASH',
    Refund: 'REFUND',
    Repayment: 'REPAYMENT',
    Collection: 'COLLECTION'
} as const;
export type TransactionTypeEnum = typeof TransactionTypeEnum[keyof typeof TransactionTypeEnum];

/**
 * @export
 */
export const TransactionStatusEnum = {
    Sent: 'SENT',
    Received: 'RECEIVED',
    Pending: 'PENDING',
    Retrying: 'RETRYING',
    Failed: 'FAILED',
    Canceled: 'CANCELED',
    Completed: 'COMPLETED',
    Reversal: 'REVERSAL',
    Refund: 'REFUND',
    Unknown: 'UNKNOWN',
    PartiallyPaid: 'PARTIALLY_PAID'
} as const;
export type TransactionStatusEnum = typeof TransactionStatusEnum[keyof typeof TransactionStatusEnum];